OTT-776: Introduce prebid analytics module for PubMatic (#442)
* OTT-776: Added basic design changes for Rejected bid Scenarios
* OTT-783: Basic design changes for pubmatic analytic module to log rejection bid scenarios (#396)
* OTT:785 Logging bids rejected due to floors, advertiser and category exclusion. (#394)
* OTT-786: Log bids rejected in adpod auction (#403)
* OTT-811: Record bid rejection metrics for analysis (#404)
* Reverting (set bid status wrt to highest(winning) combination) (#408)
* OTT-902: Remove rejected ad-pod bids from bidresponse.Ext (#424)
* OTT-904:  Introduce SeatNonBids in bidResponse.Extension (#431)
* OTT-784: As a Header bidding module, I should be able to update wrapper logger object with rejected bids (both PSB and HB side) so that analytics team can consume it
---------

+++ endpoints/openrtb2/ctv/constant/constant.go
@@ -26,8 +26,8 @@ var (
 	VASTVersionsStr = []string{"0", "1.0", "2.0", "3.0", "4.0"}
 )
 
-// BidStatus contains bids filtering reason
-type BidStatus = int
+//BidStatus contains bids filtering reason
+type BidStatus = int64
 
 const (
 	//StatusOK ...
